+++ inference_cpp/src/inference_dataset.cpp
@@ -172,10 +172,10 @@ int main(int argc, char* argv[]) {
 
         auto end_inference = std::chrono::steady_clock::now();
 
-        std::cout << "Time of invoke (ms/FPS): " << std::chrono::duration_cast<std::chrono::milliseconds>(end_invoke - start_invoke).count() << " / " << 1/(std::chrono::duration_cast<std::chrono::seconds>(end_invoke - start_invoke).count()) << std::endl;
-        std::cout << "Time of inference (ms/FPS): " << std::chrono::duration_cast<std::chrono::milliseconds>(end_inference - start_inference).count() << " / " << 1/(std::chrono::duration_cast<std::chrono::seconds>(end_inference - start_inference).count()) << std::endl;
-        ave_invoke_ms += std::chrono::duration_cast<std::chrono::milliseconds>(end_invoke - start_invoke).count();
-        ave_inference_ms += std::chrono::duration_cast<std::chrono::milliseconds>(end_inference - start_inference).count();
+        std::cout << "Time of invoke (ms/FPS): " << (float) std::chrono::duration_cast<std::chrono::milliseconds>(end_invoke - start_invoke).count() << " / " << 1/(float)(std::chrono::duration_cast<std::chrono::seconds>(end_invoke - start_invoke).count()) << std::endl;
+        std::cout << "Time of inference (ms/FPS): " << (float) std::chrono::duration_cast<std::chrono::milliseconds>(end_inference - start_inference).count() << " / " << 1/(float)(std::chrono::duration_cast<std::chrono::seconds>(end_inference - start_inference).count()) << std::endl;
+        ave_invoke_ms += (float) std::chrono::duration_cast<std::chrono::milliseconds>(end_invoke - start_invoke).count();
+        ave_inference_ms += (float) std::chrono::duration_cast<std::chrono::milliseconds>(end_inference - start_inference).count();
 
     }
